refactor : Box component marginClass 옵션을 ml,mt,mr,mb로 세분화
