+++ types/index.ts
@@ -146,13 +146,3 @@ export const TitleProps: TypographyProps = {
   type: 'mainTitle',
   text: '당신의 전적이 궁금하다면? Fizz.GG',
 };
-
-export const ErrorBoxProps: BoxProps = {
-  height: 'bgSection',
-  width: 'full',
-};
-
-export const ErrorTextPropsMapper = (error: string): TypographyProps => ({
-  type: 'title',
-  text: error,
-});
